services/scoring_24h_service/requirements.txt
confluent-kafka
redis
requests
psycopg2-binary